lib/model, lib/syncthing: move ModelWrapper to syncthing package
