+++ lib/syncthing/syncthing.go
@@ -77,7 +77,7 @@ type App struct {
 	stopOnce          sync.Once
 	mainServiceCancel context.CancelFunc
 	stopped           chan struct{}
-	Model             *model.ModelWrapper
+	Model             *ModelWrapper
 }
 
 func New(cfg config.Wrapper, dbBackend backend.Backend, evLogger events.Logger, cert tls.Certificate, opts Options) (*App, error) {
@@ -250,7 +250,7 @@ func (a *App) startup() error {
 
 	keyGen := protocol.NewKeyGenerator()
 	m := model.NewModel(a.cfg, a.myID, a.ll, protectedFiles, a.evLogger, keyGen)
-	a.Model = model.NewModelWrapper(m)
+	a.Model = newModelWrapper(m)
 
 	a.mainService.Add(m)
